Allow decimal weather hours and force 24-hour time inputs
- Weather duration fields (cerah/mendung/gerimis/hujan) now accept
  comma or dot decimals (e.g. "1,5"). Migrates weather_*_h columns
  from Int to Float and relaxes weatherSchema accordingly; totals are
  rounded to 2 decimals everywhere they're summed/displayed (outlet
  detail pages, anomaly checks, kunjungan list, CSV/XLSX export).
- Native <input type="time"> follows the browser's own locale and can
  render AM/PM regardless of any lang attribute, so jam buka/tutup
  warung now use a new TimeInput24 text field that always formats
  typed digits as 24-hour "HH:mm".

Verified via Vibium: registered an outlet with 24h times, submitted
1,5/2,25 jam weather values, and confirmed correct values through the
DB, UI totals, master edit view, and export.

+++ src/lib/validations.ts
@@ -4,12 +4,13 @@ import { hasDuplicateBrands, normalizeBrandName } from "./business-rules";
 // Sejak v1.5: total jam cuaca TIDAK wajib berjumlah/dibatasi 24 jam — interviewer bebas
 // mengisi kondisi cuaca yang benar-benar teramati tanpa harus menutupi seluruh hari. Batas
 // per-kondisi (0–24 jam masing-masing) tetap berlaku karena satu kondisi tidak mungkin lebih
-// dari 24 jam dalam sehari.
+// dari 24 jam dalam sehari. Sejak v2.6: pecahan jam (mis. 1,5 jam) diperbolehkan — angka
+// koma sudah dikonversi ke titik di sisi UI (lihat parseDecimalInput) sebelum sampai di sini.
 export const weatherSchema = z.object({
-  weatherClearH: z.coerce.number().int().min(0).max(24),
-  weatherCloudyH: z.coerce.number().int().min(0).max(24),
-  weatherDrizzleH: z.coerce.number().int().min(0).max(24),
-  weatherRainH: z.coerce.number().int().min(0).max(24),
+  weatherClearH: z.coerce.number().min(0).max(24),
+  weatherCloudyH: z.coerce.number().min(0).max(24),
+  weatherDrizzleH: z.coerce.number().min(0).max(24),
+  weatherRainH: z.coerce.number().min(0).max(24),
 });
 
 export const saleRowSchema = z.object({
